--- Sources/HostBox/IHosting.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace HostBox
-{
-    public interface IHosting
-    {
-        Task Run();
-    }
-}
