+++ Sources/HostBox/Plugins/DependencyContextExtensions.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using System.Runtime.InteropServices;
 using McMaster.NETCore.Plugins.LibraryModel;
 using Microsoft.Extensions.DependencyModel;
+using NativeLibrary = McMaster.NETCore.Plugins.LibraryModel.NativeLibrary;
 
 namespace McMaster.NETCore.Plugins.Loader
 {
