+++ Sources/HostBox/Program.cs
@@ -1,12 +1,23 @@
 ﻿using System;
+using System.IO;
+using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 
 using Common.Logging;
+using Common.Logging.Configuration;
+using HostBox.Configuration;
+using HostBox.Loading;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Configuration.Json;
+using Microsoft.Extensions.Hosting;
 
 namespace HostBox
 {
     internal class Program
     {
+        private const string ConfigurationNameEnvVariable = "configuration";
         private static ILog Logger => LogManager.GetLogger<Program>();
 
         private static async Task Main(string[] args = null)
@@ -24,10 +35,9 @@ namespace HostBox
 
                 if (commandLineArgs.CommandLineArgsValid)
                 {
-                    var host = CreateHost(commandLineArgs);
-
-                    Logger.Trace(m => m("Starting hostbox."));
-                    await host.Run();
+                    await CreateHostBuilder(commandLineArgs)
+                        .Build()
+                        .RunAsync();
                 }
             }
             catch (Exception ex)
@@ -58,14 +68,99 @@ namespace HostBox
         /// </summary>
         /// <param name="commandLineArgs"></param>
         /// <returns></returns>
-        private static IHosting CreateHost(CommandLineArgs commandLineArgs)
+        private static IHostBuilder CreateHostBuilder(CommandLineArgs commandLineArgs)
         {
+            var componentPath = Path.GetFullPath(commandLineArgs.Path, Directory.GetCurrentDirectory());
+            
+            var loader = new ComponentsLoader(
+                new ComponentConfig
+                {
+                    Path = componentPath,
+                    SharedLibraryPath = commandLineArgs.SharedLibrariesPath,
+                    LoggerFactory = LogManager.GetLogger
+                }).Load();
+            
+            var builder = new HostBuilder()
+                .ConfigureHostConfiguration(
+                    config =>
+                    {
+                        config.AddEnvironmentVariables();
+
+                        config.SetBasePath(AppDomain.CurrentDomain.BaseDirectory);
+
+                        config.AddJsonFile("hostsettings.json", true, false);
+
+                        ConfigureLogging(config.Build());
+                    })
+                .ConfigureAppConfiguration(
+                    (ctx, config) =>
+                    {
+                        LoadConfiguration(ctx.Configuration, config, componentPath, commandLineArgs);
+                    })
+                .ConfigureServices(
+                    (ctx, services) =>
+                    {
+                        loader.Run(ctx.Configuration, CancellationToken.None);
+                    });
+            
             if (commandLineArgs.Web)
             {
-                return new WebHosting(commandLineArgs);
+                var startupType = loader.EntryAssembly.GetExportedTypes()
+                    .First(t => typeof(IHostingStartup).IsAssignableFrom(t));
+                builder.ConfigureWebHost(b =>
+                {
+                    b.UseStartup<Startup>();
+                    ((IHostingStartup)Activator.CreateInstance(startupType))?.Configure(b);
+                });
             }
 
-            return new ConsoleHosting(commandLineArgs);
+            return builder;
+        }
+        
+        private static void ConfigureLogging(IConfiguration config)
+        {
+            var logConfiguration = new LogConfiguration();
+            config.GetSection("common:logging").Bind(logConfiguration);
+            LogManager.Configure(logConfiguration);
+        }
+        
+        private static void LoadConfiguration(IConfiguration currentConfiguration, IConfigurationBuilder config, string componentPath, CommandLineArgs args)
+        {
+            Logger.Trace(m => m("Loading hostable component using path [{0}].", componentPath));
+
+            var componentBasePath = Path.GetDirectoryName(componentPath);
+
+            config.SetBasePath(componentBasePath);
+
+            var configName = currentConfiguration[ConfigurationNameEnvVariable];
+
+            Logger.Info(m => m("Application was launched with configuration '{0}'.", configName));
+
+            config.LoadSharedLibraryConfigurationFiles(Logger, componentBasePath, args.SharedLibrariesPath);
+
+            var configProvider = new ConfigFileNamesProvider(configName, componentBasePath);
+
+            var templateValuesSource =
+                new JsonConfigurationSource
+                {
+                    Path = configProvider.GetTemplateValuesFile(),
+                    FileProvider = null,
+                    ReloadOnChange = false,
+                    Optional = true
+                };
+
+            templateValuesSource.ResolveFileProvider();
+
+            var templateValuesProvider = templateValuesSource.Build(config);
+
+            templateValuesProvider.Load();
+
+            foreach (var configFile in configProvider.EnumerateConfigFiles())
+            {
+                config.AddJsonTemplateFile(configFile, false, false, templateValuesProvider, args.PlaceholderPattern);
+
+                Logger.Trace(m => m("Configuration file [{0}] is loaded.", configFile));
+            }
         }
     }
 }
